don't call the inherited saveLayer in record, but just call save.
During record, our clip is wide-open (in the bug's case, 40K pixels tall). If we call saveLayer
with no bounds (as webkit does), we try to allocate an offscreen the size of the document, which
can be arbitrarily big. This cause us to run out of memory. In reality, I think there is not
benefit to calling through to saveLayer during record. On playback, we will of course call saveLayer,
but then we are on a real device (screen) with finite bounds, and the offscreen allocation is
reasonable.