update jpeg ram limits for large and small RAM class devices
http://b/issue?id=2401896

Before the fix, the image referenced in the bug would take 30+ seconds to load on a large device,
and would crash a small device. On the small device, there was no RAM cap set, since the property
we queried for does not exist there. On the large device, the properties value was too small, so
I hard-code a larger one that seems to work better.