Changes to preserve backwards compatibility for serialization
Skia's serialize methods are not intended for backwards compatibility,
but the browser used it to implement saved pages. As a result the
revision of Skia used for ICS (r1562) is not compatible with the
revision of Skia used in JB (r3000+). This CL attempts to make pages
saved in ICS render in JB while still allowing JB pages to be saved
and displayed.

bug: 6025365
Change-Id: I9983517b46d2dedf31c6d90f48f6cf00ec042bc2