Han Preference
During font initialization, create a seperate fallback list for each
locale. At runtime, use the fallbacklist associated with the locale
set in the paint object. Fallback files are associated with locales in
fallback_fonts.xml. Multiple files can be associated with the same
locale, ordering within that langauge and apart from that language in
the fallback order is preserved.

This changelist also includes some refactoring, notably of the
functions that call getNextContext().

Change-Id: Id8c91ae0be6cad8a7ef77a0cd5803676290986c1