Fixes for decoding to A8.
Do not attempt to convert non gray PNG to A8.
For an A8 image, do not consider it opaque.

cherry-picked from https://codereview.chromium.org/24882002 in Skia:

Do not convert non gray PNG to A8.

If the user requested A8, but the source is not gray, switch to
ARGB8888.

cherry-picked from https://codereview.chromium.org/25726004 in Skia:

Fixes for decoding to A8.

src/images/SkImageDecoder_libpng.cpp:
A8 images are not opaque, so do not set the opaque flag.
This fixes a bug where copyTo does not work as expected (when
copying an A8 decoded image to ARGB_8888), leading to
a bitmap hash that does not represent the image correctly
(in skimage).

tools/skimage_main.cpp:
In write_bitmap, which is creating the image for visual comparison,
copy A8 to 8888, since A8 cannot be encoded.

In the section that tests reencoding, do not test reencoding A8,
which is known to not work.

BUG:9189955