Check SkStream::isAtEnd before attempting to read.
In FrontBufferedStream, when attempting to read and buffer more
data, do not try to read if the underlying stream is at the end.

Depends on a change to frameworks base
(Idd462ff9055aea42e8c7d5d58ef177f9e667385b) to report the stream is
at the end when it throws an exception.

Prevents a bug where we continue trying to read a stream that
has already failed.

BUG:17873145
Change-Id: Ifcf9eac5da533c9966c37f6f4631251063e23ed1