Check lua into Android's Skia directory
It is already checked in to upstream, but we previously pruned it
before merging into Android. We have since changed our merge process,
which is now automated. Instead of pruning the files we take, we do a
pure git merge (plus makefile/SkUserConfig generation).

Upstream, someone modified a file in third_party/lua, which fails to
merge due to missing files here. Add the files so the merge can
succeed.

Will not affect Android build.

Change-Id: I19bbb9d67dfa3b63f9d149f1cfaa98c08430d91c