Refactor SkFILEStream for clarity.
Bug: 166069819
Test: ImageDecoderTest

Rename fSize to fEnd (since it is actually the offset of the end
underlying file, not the size of the file), fOriginalOffset to
fStart, and fOffset to fCurrent. Also initialize in the order fEnd,
fStart, then fCurrent to maintain the invariant that fCurrent should
always be kept within the range of fStart and fEnd, inclusive.

cherry-pick from https://skia-review.googlesource.com/c/skia/+/316453/3

Merged-In: I7f66558c9253d991d56fb7acf7835f514f09a571
Change-Id: I8801288320cf82b071e4621c0e770ce16016953a