Add SkFILEStream constructor which takes a size.
Bug: 166069819
Test: ImageDecoderTest

Allows the user to limit the stream to a subset of the entire file.
SkFILEStream uses the current FILE position as the base position for the
stream; this allows the user to further limit the length of the stream
to a range of bytes in the file.

cherry-pick from https://skia-review.googlesource.com/c/skia/+/316357
Change-Id: I36131ff19e75b65d501eabdfd9368f8fbb2e177f