Add back the OWNERS file to match the Skia repo
The OWNERS file was added to Skia’s repo in https://skia-review.googlesource.com/c/skia/+/443636
But it was removed from Android in https://googleplex-android-review.googlesource.com/c/platform/external/skia/+/15871448 when experimenting with how to support Skia’s OWNERS files in Android.

We now have a solution of using OWNERS.android files in b/199385287
This CL puts  back the OWNERS file to match what is in the Skia repo.

Bug: skia:12454
Change-Id: I28cf3087d71f0723a991c41711453a1c6e39d2cf