Remove libgifcodec
With b/195936927, fixed, we're now using Wuffs to decode GIF, so we no
longer build libgifcodec. While most CLs are checked into upstream Skia,
that repository no longer includes libgifcodec directly. It was merged
in from a separate repository, but with skbug.com/12902 that is no
longer the case. Remove this dead code.
Bug: skbug.com/12902
Test: make

Change-Id: Ibd3f82dec418d0fb2d87e1921b598b9c28a04f34