Change CtsEnforcement for GrGpuBufferTest
When CtsEnforcement was introduced in
Ida2c58a30c70d9a4ae0f64ec943662ba1162e406, this test was enforced at
kApiLevel_T. But this test didn't exist in T. As I understand it, we
should start enforcing the test on the version in which it was
introduced, which would be U in this case.

Update the test to only be enforced starting in Android U. Use
kNextRelease, which corresponds to the U API level, since
CtsEnforcement::kApiLevel_U does not exist in this branch.

Bug: 293367611
Test: atest CtsSkQPTestCases
Change-Id: If85169a2940af76f66b5b21d167213ea6ee910b1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/704896
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Leon Scroggins <scroggo@google.com>
Merged-In: Ie6c0c897b04a482a3aa74b9fa73421a48639230c