Fully remove SkSL error tests expecting program size limits
Follow-up to Ie4d2066b68fec925ad348f4ae675d93f865ba7ae, which attempted
to re-use the same ignore mechanism used outside of CtsSkQPTestCases.
That didn't work, so I'm just removing them entirely. As with that CL,
ProgramTooLarge_StackDepth is left as-is since it appears to still be
relevant, and isn't failing

Bug: 362646727
Bug: 362856501
Bug: 367985686
Bug: 368065469
Change-Id: I244a533f1b3e22d49aab4df6ed093cbd47e34547