Remove SK_USE_LEGACY_BLUR_GANESH staging flag
NOTE: Golden pixel tests featuring blurs may be slightly changed by this
CL. These should be rebased instead of reverting, as long as the diff is
not indicative of a buggy result.

The Ganesh-only blur code did not accurately maintain the boundary
pixels when downsampling to perform a blur at lower resolution. It had
special handling for decal that preserved transparent edges but there
was no logic for repeat or mirror tile modes. The special logic for
clamping was insufficient to prevent interior pixel content from leaking
into the boundary pixels of of the lower-resolution image.

This would then cause the interior pixel colors to bias the final blur
results near the border as the matrix kernel would accumulate more of
the boundary pixel values. This is particularly noticeable as flickering
when animating the blur radius.

The new blur algorithm code path has a rescaling approach that preserves
the original boundary colors of the source image for all tile modes so
produces much stabler blurring animations. Additionally it is also able
to skip renderpasses in complex chains of image filters, and defer
upscaling the lower-res blur output. (This doesn't apply given the
current reliance on the SkImages::MakeWithFilter API that forces scaling
back up to the final resolution, but it could be adjusted to instead
return an SkShader that automatically stretches the lowres blur to the
target view).

The new blur algorithm tends to do better at avoiding shader-based
tiling equations that can slow down the blur shader execution, so we saw
improved performance in both Graphite and Ganesh when using the new code
path.

Bug: b/40040587
Bug: b/347031058
Flag: EXEMPT third-party skia changes do not require flags
Tests: android.uirendering.cts.testclasses.RenderNodeTests#testBlurRenderEffect, android.uirendering.cts.testclasses.RenderNodeTests#testViewRenderNodeBlurEffect
Change-Id: I4cd820d19fcef207042b3a15e5f4d0f0abbf9beb